+++ src/Holding/ViewHolding/HoldingBoundWith/useBoundWithHoldings.test.js
@@ -7,10 +7,14 @@ import { renderHook, act } from '@folio/jest-config-stripes/testing-library/reac
 
 import '../../../../test/jest/__mock__';
 
-import { useOkapiKy } from '@folio/stripes/core';
-
 import { boundWithHoldingsRecords } from './fixtures';
 import useBoundWithHoldings from './useBoundWithHoldings';
+import { useTenantKy } from '../../../common';
+
+jest.mock('../../../common', () => ({
+  ...jest.requireActual('../../../common'),
+  useTenantKy: jest.fn(),
+}));
 
 const queryClient = new QueryClient();
 const wrapper = ({ children }) => (
@@ -21,7 +25,7 @@ const wrapper = ({ children }) => (
 
 describe('useBoundWithHoldings', () => {
   beforeEach(() => {
-    useOkapiKy.mockClear().mockReturnValue({
+    useTenantKy.mockClear().mockReturnValue({
       get: () => ({
         json: () => Promise.resolve({ holdingsRecords: boundWithHoldingsRecords }),
       }),
@@ -31,7 +35,7 @@ describe('useBoundWithHoldings', () => {
   it('should fetch bound-with holdings', async () => {
     const boundWithItems = [{ hrid: 'BW-ITEM-1', holdingsRecordId: '9e8dc8ce-68f3-4e75-8479-d548ce521157' }];
 
-    const { result } = renderHook(() => useBoundWithHoldings(boundWithItems), { wrapper });
+    const { result } = renderHook(() => useBoundWithHoldings(boundWithItems, 'testTenantId'), { wrapper });
 
     await act(() => !result.current.isLoading);
 
@@ -39,7 +43,7 @@ describe('useBoundWithHoldings', () => {
   });
 
   it('should not fetch bound-with holdings', async () => {
-    const { result } = renderHook(() => useBoundWithHoldings([]), { wrapper });
+    const { result } = renderHook(() => useBoundWithHoldings([], 'testTenantId'), { wrapper });
 
     expect(result.current.isLoading).toBe(false);
     expect(result.current.boundWithHoldings).toEqual([]);