+++ src/Instance/HoldingsList/Holding/HoldingContainer.js
@@ -8,6 +8,8 @@ import { withRouter } from 'react-router';
 import { Draggable } from 'react-beautiful-dnd';
 import { FormattedMessage } from 'react-intl';
 
+import { useStripes } from '@folio/stripes/core';
+
 import Holding from './Holding';
 
 const dragStyles = {
@@ -36,6 +38,9 @@ const DraggableHolding = ({
   onViewHolding,
   onAddItem,
   tenantId,
+  showViewHoldingsButton,
+  showAddItemButton,
+  isBarcodeAsHotlink,
   instanceId,
   pathToAccordionsState,
   ...rest
@@ -70,6 +75,9 @@ const DraggableHolding = ({
               onViewHolding={onViewHolding}
               onAddItem={onAddItem}
               tenantId={tenantId}
+              showViewHoldingsButton={showViewHoldingsButton}
+              showAddItemButton={showAddItemButton}
+              isBarcodeAsHotlink={isBarcodeAsHotlink}
               instanceId={instanceId}
               pathToAccordionsState={pathToAccordionsState}
             />
@@ -92,6 +100,9 @@ DraggableHolding.propTypes = {
   onViewHolding: PropTypes.func,
   onAddItem: PropTypes.func,
   tenantId: PropTypes.string,
+  showViewHoldingsButton: PropTypes.bool,
+  showAddItemButton: PropTypes.bool,
+  isBarcodeAsHotlink: PropTypes.bool,
   pathToAccordionsState: PropTypes.arrayOf(PropTypes.string),
 };
 
@@ -100,7 +111,9 @@ DraggableHolding.defaultProps = { pathToAccordionsState: [] };
 const HoldingContainer = ({
   location,
   history,
-
+  showViewHoldingsButton,
+  showAddItemButton,
+  isBarcodeAsHotlink,
   instance,
   holding,
   isDraggable,
@@ -110,10 +123,16 @@ const HoldingContainer = ({
   pathToAccordionsState,
   ...rest
 }) => {
+  const stripes = useStripes();
+
   const onViewHolding = useCallback(() => {
     history.push({
       pathname: `/inventory/view/${instance.id}/${holding.id}`,
       search: location.search,
+      state: {
+        tenantTo: tenantId,
+        tenantFrom: stripes.okapi.tenant,
+      },
     });
   }, [location.search, instance.id, holding.id]);
 
@@ -121,8 +140,12 @@ const HoldingContainer = ({
     history.push({
       pathname: `/inventory/create/${instance.id}/${holding.id}/item`,
       search: location.search,
+      state: {
+        tenantTo: tenantId,
+        tenantFrom: stripes.okapi.tenant,
+      },
     });
-  }, [instance.id, holding.id]);
+  }, [location.search, instance.id, holding.id]);
 
   return isDraggable ? (
     <Draggable
@@ -139,6 +162,9 @@ const HoldingContainer = ({
           onViewHolding={onViewHolding}
           onAddItem={onAddItem}
           tenantId={tenantId}
+          showViewHoldingsButton={showViewHoldingsButton}
+          showAddItemButton={showAddItemButton}
+          isBarcodeAsHotlink={isBarcodeAsHotlink}
           instanceId={instance?.id}
           pathToAccordionsState={pathToAccordionsState}
           {...rest}
@@ -152,6 +178,9 @@ const HoldingContainer = ({
       onViewHolding={onViewHolding}
       onAddItem={onAddItem}
       tenantId={tenantId}
+      showViewHoldingsButton={showViewHoldingsButton}
+      showAddItemButton={showAddItemButton}
+      isBarcodeAsHotlink={isBarcodeAsHotlink}
       instanceId={instance?.id}
       pathToAccordionsState={pathToAccordionsState}
     />
@@ -163,13 +192,15 @@ HoldingContainer.propTypes = {
   history: PropTypes.object.isRequired,
   provided: PropTypes.object.isRequired,
   snapshot: PropTypes.object.isRequired,
-
   instance: PropTypes.object.isRequired,
   holding: PropTypes.object.isRequired,
   holdingIndex: PropTypes.number,
   isDraggable: PropTypes.bool,
   draggingHoldingsCount: PropTypes.number,
   tenantId: PropTypes.string,
+  showViewHoldingsButton: PropTypes.bool,
+  showAddItemButton: PropTypes.bool,
+  isBarcodeAsHotlink: PropTypes.bool,
   pathToAccordionsState: PropTypes.arrayOf(PropTypes.string),
 };