+++ src/Instance/InstanceDetails/ConsortialHoldings/ConsortialHoldings.js
@@ -22,7 +22,10 @@ import {
   useSearchForShadowInstanceTenants,
 } from '../../../hooks';
 
-const ConsortialHoldings = ({ instance }) => {
+const ConsortialHoldings = ({
+  instance,
+  userTenantPermissions,
+}) => {
   const pathToAccordion = ['consortialHoldings', '_state'];
   const instanceId = instance?.id;
 
@@ -63,6 +66,7 @@ const ConsortialHoldings = ({ instance }) => {
                   key={`${memberTenant.id}.${instanceId}`}
                   memberTenant={memberTenant}
                   instance={instance}
+                  userTenantPermissions={userTenantPermissions}
                 />
               ))}
             </AccordionSet>
@@ -73,6 +77,9 @@ const ConsortialHoldings = ({ instance }) => {
   );
 };
 
-ConsortialHoldings.propTypes = { instance: PropTypes.object.isRequired };
+ConsortialHoldings.propTypes = {
+  instance: PropTypes.object.isRequired,
+  userTenantPermissions: PropTypes.arrayOf(PropTypes.object),
+};
 
 export default ConsortialHoldings;