+++ src/constants.js
@@ -631,6 +631,8 @@ export const SOURCE_VALUES = {
 
 export const CONSORTIUM_PREFIX = 'CONSORTIUM-';
 export const OKAPI_TENANT_HEADER = 'X-Okapi-Tenant';
+export const OKAPI_TOKEN_HEADER = 'X-Okapi-Token';
+export const CONTENT_TYPE_HEADER = 'Content-Type';
 
 export const DEFAULT_ITEM_TABLE_SORTBY_FIELD = 'barcode';